Add kernel package pkgrel to our package versions
The {spl,zfs}-modules-git packages depend on the Linux kernel package.
So far we haven't tracked the pkgrel of that package, meaning that on
kernel pkgrel bump we have to either bump our own pkgrel (which, by the
way, tends to get rewritten by the pkgver() function back to "1"), or
include the kernel pkgrel in our own pkgver. We implement the latter,
which is the saner approach anyway.